fix: Do not use `add_signal_handler` on Windows (#274)

+++ meltano/edk/process.py
@@ -6,6 +6,7 @@ import asyncio
 import os
 import signal
 import subprocess
+import sys
 import typing as t
 
 import structlog
@@ -135,10 +136,13 @@ class Invoker:
         )
 
         loop = asyncio.get_event_loop()
-        loop.add_signal_handler(
-            signal.SIGINT,
-            lambda s=signal.SIGINT: p.send_signal(s),  # type: ignore[misc]
-        )
+        # Windows does not support add_signal_handler
+        # https://docs.python.org/3/library/asyncio-platforms.html
+        if sys.platform != "win32":
+            loop.add_signal_handler(
+                signal.SIGINT,
+                lambda s=signal.SIGINT: p.send_signal(s),  # type: ignore[misc]
+            )
 
         streams: list[asyncio.streams.StreamReader] = []
